requirements.txt
fastapi~=0.116.2
pydantic~=2.11.9
mysql-connector-python~=9.4.0
streamlit~=1.49.1
requests~=2.32.5
pandas~=2.3.2
uvicorn~=0.35.0
pytest~=8.4.2
